# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StoryWeaver",
    "description": "Production-grade FastAPI backend with auth (JWT+OAuth+RBAC) and LLM chat proxy domains.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/works/{work_id}/scenes/{scene_id}/assist/continue": {
      "post": {
        "tags": [
          "assist"
        ],
        "summary": "이어쓰기 (SSE)",
        "operationId": "assist_continue_api_v1_works__work_id__scenes__scene_id__assist_continue_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContinueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/works/{work_id}/scenes/{scene_id}/assist/infill": {
      "post": {
        "tags": [
          "assist"
        ],
        "summary": "인필링 (SSE)",
        "operationId": "assist_infill_api_v1_works__work_id__scenes__scene_id__assist_infill_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InfillRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/works/{work_id}/scenes/{scene_id}/assist/dialogue": {
      "post": {
        "tags": [
          "assist"
        ],
        "summary": "지문/대사 변환 (SSE)",
        "operationId": "assist_dialogue_api_v1_works__work_id__scenes__scene_id__assist_dialogue_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DialogueRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/works/{work_id}/scenes/{scene_id}/assist/style": {
      "post": {
        "tags": [
          "assist"
        ],
        "summary": "문체 변환 (SSE)",
        "operationId": "assist_style_api_v1_works__work_id__scenes__scene_id__assist_style_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StyleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/works/{work_id}/scenes/{scene_id}/assist/correct": {
      "post": {
        "tags": [
          "assist"
        ],
        "summary": "교정 (SSE)",
        "operationId": "assist_correct_api_v1_works__work_id__scenes__scene_id__assist_correct_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "work_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Work Id"
            }
          },
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CorrectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContinueRequest": {
        "properties": {
          "cursorText": {
            "type": "string",
            "title": "Cursortext"
          }
        },
        "type": "object",
        "required": [
          "cursorText"
        ],
        "title": "ContinueRequest"
      },
...
      "CorrectRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CorrectRequest"
      },
      "DialogueRequest": {
        "properties": {
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "targetEntityId": {
            "type": "string",
            "format": "uuid",
            "title": "Targetentityid"
          }
        },
        "type": "object",
        "required": [
          "intent",
          "targetEntityId"
        ],
        "title": "DialogueRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InfillRequest": {
        "properties": {
          "beforeText": {
            "type": "string",
            "title": "Beforetext"
          },
          "afterText": {
            "type": "string",
            "title": "Aftertext"
          }
        },
        "type": "object",
        "required": [
          "beforeText",
          "afterText"
        ],
        "title": "InfillRequest"
      },
...
      "StyleRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "targetStyle": {
            "type": "string",
            "title": "Targetstyle"
          }
        },
        "type": "object",
        "required": [
          "text",
          "targetStyle"
        ],
        "title": "StyleRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}